requirements.txt
blinker==1.4
brotlipy==0.7.0
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
opencv-python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pyparsing==3.0.9
streamlit==1.23.1
sympy==1.12
tenacity==8.2.2
torch==2.0.1
torchvision==0.15.2

